+++ spec/system/task_spec.rb
@@ -1,21 +1,8 @@
 require 'rails_helper'
 
-def retry_on_stale_element_reference_error(&block)
-  first_try = true
-  begin
-    block.call
-  rescue Selenium::WebDriver::Error::StaleElementReferenceError
-    warn 'Retry on StaleElementReferenceError'
-    if first_try
-      first_try = false
-      retry
-    end
-    raise
-  end
-end
-
 RSpec.describe 'Tasks', type: :system do
   include CreateTestTasksSupport
+  include RetryTestSupport
   describe 'タスクの作成' do
     let(:today) { Time.zone.today }
     let!(:category) { create(:category) }
@@ -168,8 +155,6 @@ RSpec.describe 'Tasks', type: :system do
 
         # もう一度押すと降順に並べ替えられる
         click_on '重要度'
-
-        # TODO find css 赤文字のcss
 
         retry_on_stale_element_reference_error do
           tasks = page.all('.task')
